fix: code fixes and renaming
